fix(monitor): align dashboard with VPS ground truth (Xray ghost count, vnstat window, speed precision, auto-sync)
Root-cause analysis on a live deploy uncovered four independent bugs that
made the dashboard's numbers diverge from the VPS panel by 100%+ on
Active Users and silently freeze whenever no external cron was running.

Bugs and fixes:

1. Xray ghost accounts (PRIMARY — caused 'Active User: 70' vs panel '0')
   The agent counted /etc/xray/.userall.db lines as 'lifetime active'
   when they had no parseable expiry date. Real Premium installer
   deployments accumulate dateless entries in this file for every
   account ever created (including ones already deleted from the panel),
   so the count drifted upward forever. The audit case study had 41
   such ghost lines pushing 27 real SSH subscribers up to a reported
   68 active_users.

   Drop /etc/xray/.userall.db from the agent's data sources entirely.
   Read /usr/local/etc/xray/config.json (the live config Xray itself
   uses) and count unique 'email' entries — the same source the panel
   reads. Return 0 when xray.service is inactive, mirroring panel
   behavior. _count_active_ssh_in_db now also skips dateless lines
   instead of treating them as lifetime.

2. vnstat lifetime vs monthly window
   Agent used interfaces[0].traffic.total which is lifetime cumulative
   on whichever interface vnstat happened to list first (could be lo,
   tun0, etc on multi-NIC hosts). Now picks the kernel default-route
   interface explicitly and reads traffic.month[<latest>] so figures
   match what 'vnstat -m' shows on the box. Falls back to lifetime
   total when no monthly bucket exists, then to psutil since-boot.

3. Speed truncation to 0 for sub-Mbps traffic
   _throughput_mbps did int(bits/1e6) — every traffic level under
   1 Mbps truncated to 0 and the dashboard rendered '—'. Operators
   thought the agent was broken when in fact a VPN with 0.4 Mbps avg
   was working fine. Agent now returns float (1 decimal). Schema
   speedMbps is Float; new formatSpeed() utility renders <10 Mbps with
   1 decimal, >=10 Mbps as integer, 0/null/non-finite as '—'.
   Zod validators on /api/servers updated to allow non-int speed.

4. No internal scheduler (data frozen forever without external cron)
   /api/monitor/sync only ran from admin button or curl-based crontab.
   When the operator forgot to install the cron, dashboard stayed
   stuck at the last manual sync — chart history went flat, RX/TX
   never updated, ping appeared static. The audit found crontab -l
   completely empty on the deploy under investigation.

   Add maybeAutoSync() in lib/monitor: fire-and-forget background
   trigger called from /api/servers/public and /api/stats GET handlers.
   Singleton-guarded (only one inflight at a time), 30 s cooldown
   between successive runs, skips entirely when no server is stale.
   Tunable via MONITOR_AUTOSYNC_COOLDOWN_MS / MONITOR_AUTOSYNC_STALE_MS;
   kill switch MONITOR_AUTOSYNC_DISABLED=1. External cron still
   recommended in DEPLOY.md as a safety net for low-traffic deploys.

Other cleanup:
- Drop unused 'online_now' field from agent payload (dashboard never
  consumed it; it was silent drift in the contract).
- Update DEPLOY.md: section 6.2 documents agent reinstall flow,
  section 9 explains the new auto-sync, troubleshooting matrix gains
  three v1.1 migration entries.

Verified:
- python3 -m py_compile on the agent: OK
- tsc --noEmit: clean
- prisma generate: OK on updated schema (Int->Float on speedMbps is
  a non-destructive change for SQLite via prisma db push)

+++ src/app/api/servers/[id]/route.ts
@@ -38,7 +38,7 @@ const updateSchema = z.object({
   status: z.enum(["ONLINE", "OFFLINE", "FULL", "WARNING", "UNKNOWN"]).optional(),
   activeUsers: z.number().int().min(0).optional(),
   pingMs: z.number().int().min(0).optional(),
-  speedMbps: z.number().int().min(0).optional(),
+  speedMbps: z.number().min(0).optional(),
 });
 
 export const dynamic = "force-dynamic";
