+++ src/app/api/servers/route.ts
@@ -39,7 +39,7 @@ const createSchema = z.object({
   // optional manual override
   activeUsers: z.number().int().min(0).optional(),
   pingMs: z.number().int().min(0).optional(),
-  speedMbps: z.number().int().min(0).optional(),
+  speedMbps: z.number().min(0).optional(),
 });
 
 export const dynamic = "force-dynamic";
